+++ mordant/src/linuxX64Main/kotlin/com/github/ajalt/mordant/internal/MppImpl.kt
@@ -4,8 +4,7 @@ import kotlinx.cinterop.alloc
 import kotlinx.cinterop.memScoped
 import platform.posix.*
 
-@OptIn(ExperimentalUnsignedTypes::class)
-internal actual fun getTerminalSize(timeoutMs: Long): Pair<Int, Int>? {
+internal actual fun getTerminalSize(): Pair<Int, Int>? {
     return memScoped {
         val size = alloc<winsize>()
         if (ioctl(STDIN_FILENO, TIOCGWINSZ, size) < 0) {
@@ -15,5 +14,3 @@ internal actual fun getTerminalSize(timeoutMs: Long): Pair<Int, Int>? {
         }
     }
 }
-
-internal actual fun isWindows(): Boolean = false