+++ mordant/src/mingwX64Main/kotlin/com/github/ajalt/mordant/internal/MppImpl.kt
@@ -5,7 +5,7 @@ import platform.windows.*
 
 
 // https://docs.microsoft.com/en-us/windows/console/getconsolescreenbufferinfo
-internal actual fun getTerminalSize(timeoutMs: Long): Pair<Int, Int>? = memScoped {
+internal actual fun getTerminalSize(): Pair<Int, Int>? = memScoped {
     val csbi = alloc<CONSOLE_SCREEN_BUFFER_INFO>()
     val stdoutHandle = GetStdHandle(STD_OUTPUT_HANDLE)
     if (stdoutHandle == INVALID_HANDLE_VALUE) {
@@ -15,10 +15,8 @@ internal actual fun getTerminalSize(timeoutMs: Long): Pair<Int, Int>? = memScope
     if (GetConsoleScreenBufferInfo(stdoutHandle, csbi.ptr) == 0) {
         return@memScoped null
     }
-    csbi.dwSize.run { X.toInt() to Y.toInt() }
+    csbi.srWindow.run { Right - Left + 1 to Bottom - Top + 1 }
 }
-
-internal actual fun isWindows(): Boolean = true
 
 // https://docs.microsoft.com/en-us/windows/console/setconsolemode
 // https://docs.microsoft.com/en-us/windows/console/getconsolemode